Import and export options (#31)
* Implement import and export of movies

Use existing tableView instance (+66 squashed commits)
Squashed commits:
[8a20ae5] Fix irregular failing test with hidden tableView
[ec22f69] Customize some alert messages
[0146b97] Refactor code for tests
[07ec642] Reset code signing settings
[ed9e19f] Add completionHandler for resetting database before import
[c62bfb3] Remove temporarily setting list position with counter

This is not used in iOS yet and android app can handle a
list of movies with listPosition of zero, so a temporary
workaround is not needed at the moment.
[e35aab6] Rename export file for android compatility
[1a54a3c] Only encode values if present
[114b37e] Use MovieStorage to clean up database before import
[bdd24d9] Explicitly define a completion handler

Define defaultActionHandler eyplicitly because there
are two handler available.
[a4266c0] Remove todos as they won't be part of this feature
[7038dcc] Handle error when export failed
[c9399db] Set initial value of listPosition to 0
[c612323] Add formatter for runtime and voteAverage

Display voteAverage always with one digital number or
display a placeholder value to show that there is no data.
[951ee3e] Add listPosition to storedMovie entities
[737feb0] Sort after listPosition and title
[52d4c58] Set development language to german
[ecc75b3] Clean up code
[206dd97] Make voteAverage and runtime of Movie optional

This is due to some changes of movieDB api where movies had no voteAverage and runtime key.
[03ea1d0] Check in IDEWorkspaceChecks
[b3ef1b7] Update xcode project to recommended settings
[5b55771] Create OrangeNavigationController

Instead of customizing all navigation controllers with UIAppearance,
use a subclass for internal navigationController to resolve problems
with different designs of viewControllers outside the apps sandbox
like UIDocumentInteractionController and UIDocumentPickerViewController.
[8b4cfe4] Load all posters after importing movies
[b200f60] Set locale of dateFormatter to US
[f30f25c] Fix tests
[f9bcf9b] Inject storageManager as dependency
[8eff4df] Clean up some code
[863d0a9] Make StoredMovie conform to Decodable protocol
[7a31839] Fix test
[21fdf84] Save instance of UIDocumentInteractionController

This is neccessary to not loose the reference to the controller
and to successfully export data to another app.
[32bd998] Set voteAverage to type Decimal

This will solve the problem with the number of decimals when encoding
the model to json. With the type of Decimal this will be done correctly.
[c2f1fb7] Make StoredMovie conform to encodable protocol

Use JSONEncoder to encode all storedMovies.
[a649386] Add list position to StoredMovie
[8d44c12] Remove old data on import
[eaaf0f6] Present warning for user before the file is imported

Inform the user, that his database will be overwritten, if the
database is not empty. This information is displayed before the
user is navigating to the file which should be imported.
[3754f1d] Present a modal viewcontroller while importing movies
[2884b2f] Only export data when database is not empty
[6fb95ca] Remove xcshareddata folder from gitignore
[c4f8cd5] Add two todos
[d2a0e6a] Add string for uti
[4883a5c] Remove class variable of docController
[dfadcea] Ask user to import data if database is not empty
[69dd484] Setup fetchedResultManager in seperate function
[9e88bb6] Remove cancelling delegate as it won't be handled
[4e1e549] Add import messages for alert
[a65ef6e] Style status bar in settingsVC
[76b523b] Add import feature with UIDocumentPickerVC
[813e773] Add exportMoviesFileName to Strings
[0161591] Add exception handling
[0375bcc] Remove preview of file and uti type
[5933e2c] Use UIDocumentInteractionController to share json
[0d4321a] Prepare export in settingsViewController
[d66050d] Make completionHandler optional

The completionHandler for setting up and updating the fetchedResultsManager
is now optional and has a default with nil, so we must not call an empty
completion block when we do nothing on completion
[8fe4b88] Remove automatic export in MoviesVC
[7ccba8c] Update predicate to load all stored movies for export
[c33b140] Truncate file at offset 0

This will delete the data from file and fill it again with data.
Don't know if this is a good solution but it works atm.
[e8d41ac] Add possibility to load generated json from documentsDirectory again
[4e9182b] Add possibility to save generated json to documentsDirectory
[c5a6172] Rename function
[7da2d41] Add voteCount for creating StoredMovie in test
[e4edfc6] Add xcshareddata to gitignore for Xcode 9.3 beta
[9ab453e] Add call for hiding table view to viewDidAppear
[46bc9f5] Call the function somewhere for testing purpose
[c843c36] Add first exportData function

This function will convert the core data model to json
[fcad124] Add voteCount and watchedDate to core data model
[2f76811] Add date formatter for export json

* Fix design in search and status bar

* Handle decoding optional releaseDate key

Set releaseDate in Movie model to optional.
Add empty state strings for releaseDate, runtime, voteAverage and voteCount.

* Use the medium dateStyle for formatted dates

This will automatically adapt the dateStyle for the current locale.

* Sort keys by usage in initializer

* Make loadPoster function in Movie class static

This will reduce redundancy in code in StoredMovie

* Directly save imported sotredMovies to the context

* Add ImportExportObject to add an extra key to json

The array of movies have to be nested in a movies object key to
get a json that is equal to the json of the android app. They
already have series and movies, so an extra key is necessary.
We ignore the series key at the moment.

* Delete CodingKeys because the keys are not neccessary

* Move reset CoreData to MovieStorage