Change target and scheme names
The dev target/scheme is now called "Cineaste App-Dev".
The appStore target/scheme is now called "Cineaste App".
The reason for this is the text in an unconfigurable permission
alert for nearby. In this alert the name of the target is used
and it feels much better (for the AppStore version) when the name
of the app you grant access to is called "Cineaste App" instead of
"Cineaste-AppStore"