Don't show region information for release date
For the `discover` and `search` requests, we do not know if the
localized or the primary release date is returned, so we can't
display any information about this. For the `movie` detail request,
it would be possible, but therefore we need another property to
say if it is the localized one or not. This is not really worth
the effort. The result of this is, that we don't save any
localizedReleaseDate information at all, if we can't even
differentiate if it is localized or not.