Use the localized release date (#43)
* Add localizedReleaseDate in movie requests

* Update requests to use the localized release date and not
the primary release date
* Add localizedReleaseDate to StoredMovie CoreData model
* Split files of Movie and StoredData

* Display region identifier after release date

Identifier will only displayed in movieDetail and not in the lists

* Let localizedReleaseDate decoding not throw error

When there is an error on decoding the localizedReleaseDates response,
this should not throw an error, because the localizedReleaseDate is an
optional value and the user should not see an alert in an error case.

* Fix tests

* Change TEST_TARGET_NAME for UITests

* Update target membership for added files

Run `pod install`

* Simplify decoding of LocalizedReleaseDate

* Don't show region information for release date

For the `discover` and `search` requests, we do not know if the
localized or the primary release date is returned, so we can't
display any information about this. For the `movie` detail request,
it would be possible, but therefore we need another property to
say if it is the localized one or not. This is not really worth
the effort. The result of this is, that we don't save any
localizedReleaseDate information at all, if we can't even
differentiate if it is localized or not.

* Only use type 3 of releaseDate

* Update target membership for SafariViewController

* Set swiftlint rules to strict again

* Use correct language for localized release date

* Fix UITest