Show "Swipe Action" Hint only three times (#104)
* #98 Move showing of the "Swipe action" hint to an extension

This allows for additional logic to be in one file,
encapsulating it to improve readability.

* #98 Use `guard` statement for precondition

The fact that a cell exists is a precondition to animating
the hint, which can be made even more obvious with `guard`.

* #98 Add documentation

* #98 Only show "Swipe action" hint once

The number of times that the hint should be shown
can be configured using the constant.

* #98 Set number of times to show the swipe action hint to 3

As per @fbernutz's suggestion. This is a configuration
change and can be adjusted if needed.

* #98 Add enum for persisting the "swipe action" show count

This keeps the access to `UserDefaults` persistent within
the other persistency enums in the code base.

* #98 Make use of the `SwipeActionHintShowCountPersistence`

By using this enum, we're making sure that accessing
`UserDefaults` is consistent across the code base.